backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
alembic==1.13.2
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic[email]==2.8.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
httpx==0.27.0
anthropic==0.30.0
boto3==1.34.144
redis==5.0.7
celery==5.4.0
PyMuPDF==1.24.7
pytesseract==0.3.13
Pillow==10.4.0
openpyxl==3.1.5
pytest==8.2.2
pytest-asyncio==0.23.7